modify manage.database and add token_manage
